feat : 코드 초기화 및 요구사항 분석
