--- src/main/java/Application.java
@@ -1,8 +1,0 @@
-import Controller.Controller;
-
-public class Application {
-    public static void main(String[] args) {
-        Controller controller = new Controller();
-        controller.gamePlay();
-    }
-}
