+++ src/main/java/racingcar/Application.java
@@ -1,0 +1,6 @@
+package racingcar;
+
+public class Application {
+    public static void main(String[] args) {
+    }
+}
