+++ src/main/java/racingcar/controller/Controller.java
@@ -1,0 +1,5 @@
+package racingcar.controller;
+
+public class Controller{
+
+}
