+++ src/main/java/racingcar/message/ConsoleMessage.java
@@ -1,4 +1,4 @@
-package Message;
+package racingcar.message;
 
 public enum ConsoleMessage {
     INPUT_CAR_NAME("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분)."),
