+++ src/main/java/racingcar/model/Car.java
@@ -1,0 +1,5 @@
+package racingcar.model;
+
+public class Car {
+
+}
