+++ src/main/java/racingcar/model/Name.java
@@ -1,0 +1,7 @@
+package racingcar.model;
+
+import java.util.ArrayList;
+
+public class Name {
+
+}
