+++ src/main/java/racingcar/view/Input.java
@@ -1,0 +1,7 @@
+package racingcar.view;
+
+import java.util.Scanner;
+
+public class Input {
+
+}
