+++ src/main/java/racingcar/view/Output.java
@@ -1,0 +1,11 @@
+package racingcar.view;
+
+import racingcar.model.Car;
+
+import java.util.ArrayList;
+
+import static racingcar.message.ConsoleMessage.OUTPUT_WINNER;
+
+public class Output {
+
+}
