--- src/test/java/racingcar/ApplicationTest.java
@@ -1,25 +1,0 @@
-package racingcar;
-
-import Model.Car;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
-public class ApplicationTest {
-
-    @Test
-    @DisplayName("이름이 5글자 초과면 예외")
-    void nameLengthEx(){
-        assertThatThrownBy(() -> new Car("longName"))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    /*@Test
-    @DisplayName("이름이 잘못들어오면 예외")
-    void nameValidEx(){
-        assertThatThrownBy(() -> namesc
-
-    }*/
-}
